refactor: add pipeline runner service
