+++ app/views/orchestration/actions/_form.html.erb
@@ -25,7 +25,7 @@
 
   <div>
     <%= f.label :agent_class, class: "block text-sm font-medium text-gray-700 mb-1" %>
-    <%= f.text_field :agent_class, class: "w-full border border-gray-300 rounded-lg px-3 py-2 text-sm font-mono focus:outline-none focus:ring-2 focus:ring-indigo-500 #{"border-red-400" if action.errors[:agent_class].any?}", placeholder: "e.g. EmailClassifyAgent" %>
+    <%= f.text_field :agent_class, class: "w-full border border-gray-300 rounded-lg px-3 py-2 text-sm font-mono focus:outline-none focus:ring-2 focus:ring-indigo-500 #{"border-red-400" if action.errors[:agent_class].any?}", placeholder: "e.g. Emails::ClassifyAgent" %>
     <% if action.errors[:agent_class].any? %>
       <p class="mt-1 text-xs text-red-600"><%= action.errors[:agent_class].first %></p>
     <% end %>
